get data from api is successfully implemented

+++ app/src/main/java/com/poepoemyintswe/foodmagnet/model/Location.java
@@ -4,6 +4,6 @@ package com.poepoemyintswe.foodmagnet.model;
  * Created by poepoe on 2/4/15.
  */
 public class Location {
-  public Float lat;
-  public Float lng;
+  public double lat;
+  public double lng;
 }
